Moved the application start procedures to its on module so it can be tested.
